+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# 1.3.0 [unreleased]
+# 1.3.1 [11/26/2022]
+
+## 🐛 Bugs 🐛
+
+-   **Fixed** issues with the slider styling in firefox [#73](https://github.com/EricTurf/t-components/pull/73)
+
+# 1.3.0 [11/26/2022]
 
 ## ✨Enhancement ✨